refactor: combine mocks and remove stale ref
