+++ packages/salesforcedx-utils-vscode/test/unit/commands/channelService.test.ts
@@ -6,7 +6,6 @@
  */
 
 import { expect } from 'chai';
-import { EOL } from 'os';
 import * as proxyquire from 'proxyquire';
 import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
 import {
@@ -15,35 +14,7 @@ import {
   SfdxCommandBuilder
 } from '../../../src/cli';
 import { nls } from '../../../src/messages';
-
-class MockChannel {
-  public readonly name = 'MockChannel';
-  public value = '';
-
-  public append(value: string): void {
-    this.value += value;
-  }
-
-  public appendLine(value: string): void {
-    this.value += value;
-    this.value += EOL;
-  }
-
-  public clear(): void { }
-  public show(preserveFocus?: boolean | undefined): void;
-  public show(column?: any, preserveFocus?: any) { }
-  public hide(): void { }
-  public dispose(): void { }
-}
-
-const vscodeStub = {
-  window: {
-    // @ts-ignore
-    createOutputChannel: mockChannel => {
-      return mockChannel;
-    }
-  }
-};
+import { MockChannel, vscodeStub } from './mocks';
 
 const { ChannelService } = proxyquire.noCallThru()('../../../src/commands', {
   vscode: vscodeStub
